requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database - Versi terbaru yang support Python 3.13
sqlalchemy==2.0.36
pymysql==1.1.1

# Config
python-dotenv==1.0.1

# Scheduler
apscheduler==3.10.4

# Templates & Static
jinja2==3.1.4

# File Upload & Processing
python-multipart==0.0.12
pandas==2.2.3
openpyxl==3.1.5

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Database Migration
alembic==1.14.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.0

# Caching
redis==7.1.0
